+++ lib/Traits/Trait.ts
@@ -1,5 +1,19 @@
+import { JsonObject } from "../Core/Json";
 import Result from "../Core/Result";
 import { BaseModel } from "../Models/Model";
+
+export interface TraitClassOptions {
+  description?: string;
+  example?: JsonObject;
+}
+
+/** Decorator to set traitClass options (eg `description` of the class) */
+export function traitClass(options: TraitClassOptions) {
+  return function(target: any) {
+    target.description = options.description;
+    target.example = options.example;
+  };
+}
 
 export interface TraitOptions {
   name: string;
@@ -7,6 +21,9 @@ export interface TraitOptions {
 }
 
 export default abstract class Trait {
+  static description: string | undefined;
+  static example: string | undefined;
+
   readonly id: string;
   readonly name: string;
   readonly description: string;