+++ lib/Traits/TraitsClasses/WebMapServiceCatalogItemTraits.ts
@@ -6,6 +6,7 @@ import primitiveArrayTrait from "../Decorators/primitiveArrayTrait";
 import primitiveTrait from "../Decorators/primitiveTrait";
 import mixTraits from "../mixTraits";
 import ModelTraits from "../ModelTraits";
+import { traitClass } from "../Trait";
 import CatalogMemberTraits from "./CatalogMemberTraits";
 import DiffableTraits from "./DiffableTraits";
 import ExportableTraits from "./ExportableTraits";
@@ -21,17 +22,6 @@ import UrlTraits from "./UrlTraits";
 export const SUPPORTED_CRS_3857 = ["EPSG:3857", "EPSG:900913"];
 export const SUPPORTED_CRS_4326 = ["EPSG:4326", "CRS:84", "EPSG:4283"];
 
-/**
- * Creates a single item in the catalog from one or many WMS layers.<br/>
- * <strong>Note:</strong> <i>To present all layers in an available WMS as individual items in the catalog use the \`WebMapServiceCatalogGroup\`.</i>
- * @example
- * {
- *   "type": "wms",
- *   "name": "Mangrove Cover",
- *   "url": "https://ows.services.dea.ga.gov.au",
- *   "layers": "mangrove_cover_v2_0_2"
- * }
- */
 export class WebMapServiceAvailableStyleTraits extends ModelTraits {
   @primitiveTrait({
     type: "string",
@@ -147,6 +137,17 @@ export class WebMapServiceAvailableLayerDimensionsTraits extends ModelTraits {
   dimensions?: WebMapServiceAvailableDimensionTraits[];
 }
 
+@traitClass({
+  description: `Creates a single item in the catalog from one or many WMS layers.
+
+<strong>Note:</strong> <i>To present all layers in an available WMS as individual items in the catalog use the \`WebMapServiceCatalogGroup\`.</i>`,
+  example: {
+    type: "wms",
+    name: "Mangrove Cover",
+    url: "https://ows.services.dea.ga.gov.au",
+    layers: "mangrove_cover_v2_0_2"
+  }
+})
 export default class WebMapServiceCatalogItemTraits extends mixTraits(
   ExportableTraits,
   DiffableTraits,